Created controller and template for handling requests to the annotator.
Created controller and template for handling requests to the annotator. These are used for requests made to '/' and '/annotator'.
Both these routes have been used for the annotator as there is no need for a dedicated homepage to the project but this future proofs the '/annotator' route if the '/' route is changed.