Removed unnecessary content from the base template.
Removed unnecessary content from the base template and created a miscellaneous template and renamed the BaseController to MiscController. This was done as not every page needs to use a specific controller and a custom template. Therefore the misc template takes an HTML string as the content and as a result can display anything required. This also allows it to be used to view the changes that are made to the base template as it does not override anything that the base template sets up.